most of the ins file functionality is detached to allow other file formats being edited in the ins form; the dispersion DISP card is treated in similar way to extended SFAC card now and is a part of the RefinementModel; TODO: provide an interface to collect all these scattered data to a single scatterer object... Persistence is in better condition now - all atoms but deleted are saved now, thus preserving the model state. A crash is fixed in ConnInfo::Compile; PersistentIds as opposite to normal Ids are introduced for the persistence model...

+++ trunk/gxlib/gllabels.cpp
@@ -77,7 +77,7 @@ bool TXGlLabels::Orient(TGlPrimitive& P)  {
     if( Mode & lmPart )  {
       if( ca.GetPart() != 0)  {
         if( Tmp.Length() )  Tmp << ", ";
-        Tmp << ca.GetPart();
+        Tmp << (int)ca.GetPart();
       }
     }
     if( Mode & lmAfix )  {
